Bug 1811261 - Part 1: Move address autofill related UI tests to new class
