+++ app/src/androidTest/java/org/mozilla/fenix/ui/SettingsBasicsTest.kt
@@ -121,71 +121,6 @@ class SettingsBasicsTest {
 
     @SmokeTest
     @Test
-    fun verifyAddressAutofillTest() {
-        val addressFormPage =
-            TestAssetHelper.getAddressFormAsset(mockWebServer)
-
-        homeScreen {
-        }.openThreeDotMenu {
-        }.openSettings {
-        }.openAutofillSubMenu {
-            clickAddAddressButton()
-            fillAndSaveAddress(
-                "Mozilla",
-                "Fenix",
-                "Firefox",
-                "Harrison Street",
-                "San Francisco",
-                "Alaska",
-                "94105",
-                "United States",
-                "555-5555",
-                "foo@bar.com",
-            )
-        }.goBack {
-        }.goBack {
-        }
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(addressFormPage.url) {
-            clickStreetAddressTextBox()
-            clickSelectAddressButton()
-            clickAddressSuggestion("Harrison Street")
-            verifyAutofilledAddress("Harrison Street")
-        }
-    }
-
-    @SmokeTest
-    @Test
-    fun deleteSavedAddressTest() {
-        homeScreen {
-        }.openThreeDotMenu {
-        }.openSettings {
-        }.openAutofillSubMenu {
-            clickAddAddressButton()
-            fillAndSaveAddress(
-                "Mozilla",
-                "Fenix",
-                "Firefox",
-                "Harrison Street",
-                "San Francisco",
-                "Alaska",
-                "94105",
-                "United States",
-                "555-5555",
-                "foo@bar.com",
-            )
-            clickManageAddressesButton()
-            clickSavedAddress("Mozilla")
-            clickDeleteAddressButton()
-            clickCancelDeleteAddressButton()
-            clickDeleteAddressButton()
-            clickConfirmDeleteAddressButton()
-            verifyAddAddressButton()
-        }
-    }
-
-    @SmokeTest
-    @Test
     fun verifyCreditCardAutofillTest() {
         val creditCardFormPage = TestAssetHelper.getCreditCardFormAsset(mockWebServer)
